+++ web/app/api/auth/[...nextauth]/route.ts
@@ -59,7 +59,7 @@ const handler = NextAuth({
     },
   },
   pages: {
-    signIn: "/auth/signin",
+    signIn: "/",
     error: "/auth/error",
   },
   session: {
